Corrige un crash au login : downloadSpeed n'est pas toujours un entier
Plantage systématique juste après connexion sur la vraie prod :
JsonDataException "Expected a long but was 7485.6" sur
$.items[0].download.downloadSpeed. Le moteur BitTorrent renvoie un
débit brut non arrondi (voir t.downloadSpeed dans
src/app/api/activity/v2/route.ts, contrairement à eta qui passe par
Math.round côté serveur) — QueueDownloadDto.downloadSpeed déclaré en
Long ne pouvait pas parser une valeur décimale. Passé en Double,
confirmé contre la vraie file de téléchargement en prod (débits
7554.8, 17500.2...).

+++ android-tv/app/src/main/kotlin/com/movviz/tv/data/ApiModels.kt
@@ -267,7 +267,13 @@ data class QueueDownloadDto(
     // useSmoothProgress/DownloadQueue.tsx côté desktop qui multiplie par 100
     // à l'affichage seulement.
     val progress: Double = 0.0,
-    val downloadSpeed: Long = 0,
+    // Double, pas Long : le moteur BitTorrent renvoie un débit brut non
+    // arrondi (voir t.downloadSpeed dans src/app/api/activity/v2/route.ts,
+    // contrairement à `eta` qui lui passe par Math.round côté serveur) —
+    // un Long ici plantait le parsing JSON dès qu'un torrent avait un débit
+    // non entier (crash confirmé en direct juste après connexion : "Expected
+    // a long but was 7485.6").
+    val downloadSpeed: Double = 0.0,
     val eta: Long = 0, // secondes
     val state: String = "downloading",
 )
